edge/ai-service/requirements.txt
# Python dependencies for Edge AI Service
# Core dependencies
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.9.0
grpcio>=1.66.0
grpcio-tools>=1.66.0

# AI/ML dependencies
openvino>=2024.0.0
openvino[tools]>=2024.0.0  # Includes Model Optimizer for conversion
onnxruntime>=1.20.0
numpy>=1.26.0
opencv-python>=4.10.0
pillow>=10.4.0
ultralytics>=8.3.0  # For YOLOv8 model download and conversion
torch>=2.0.0  # Required by ultralytics

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Development dependencies
pytest>=8.3.0
pytest-asyncio>=0.24.0
pytest-cov>=5.0.0
pytest-mock>=3.14.0
httpx>=0.27.0  # For testing FastAPI endpoints
# black>=24.10.0
# pylint>=3.2.0

